requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cycler==0.11.0
Django==4.1
django-filter==22.1
django-ipware==4.0.2
django-rest-swagger==2.2.0
django-user-agents==0.4.0
djangorestframework==3.13.1
docopt==0.6.2
drf-yasg==1.21.3
fonttools==4.37.1
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.1
numpy==1.23.2
openapi-codec==1.3.2
packaging==21.3
Pillow==9.2.0
pipreqs==0.4.11
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
ua-parser==0.16.1
uritemplate==4.1.1
urllib3==1.26.12
user-agents==2.2.0
yarg==0.1.9
